# docs/openapi_agent_schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EcoMind-AI 环保管家智能体工具 API",
    "description": "EcoMind-AI 智慧环保平台开放接口，供 AI 智能体调用。支持设备状态查询、实时监测数据、报警管理、AI 趋势预测和运维诊断报告。",
    "version": "1.0.0",
    "contact": {
      "name": "EcoMind-AI"
    }
  },
  "servers": [
    {
      "url": "https://ecomind-ai-193380-4-1304218020.sh.run.tcloudbase.com",
      "description": "生产环境（腾讯云 CloudBase 云托管）"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "EcoMind-AI 开放平台 API Key，通过管理后台创建。支持 single_org（绑定企业）和 all_orgs（全企业查询，调用时需传 enterprise_name/org_id）两种模式。"
      }
    }
  },
  "paths": {
...
    "/openapi/alarm/acknowledge": {
      "post": {
        "operationId": "acknowledge_alarm",
        "summary": "确认报警",
        "description": "确认（处理）一条报警记录，将其状态从'待处理'更新为'已确认'。single_org Key 默认查绑定企业；all_orgs Key 需传 enterprise_name 或 org_id 指定企业。",
        "parameters": [
          {
            "name": "enterprise_name",
            "in": "query",
            "required": false,
            "description": "企业名称（all_orgs Key 必填，支持模糊匹配）",
            "schema": { "type": "string" }
          },
          {
            "name": "org_id",
            "in": "query",
            "required": false,
            "description": "组织ID（可选，优先于 enterprise_name）",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "alarm_id",
            "in": "query",
            "required": true,
            "description": "报警记录ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "报警确认结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "summary": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "alarm_id": { "type": "string" },
                        "new_status": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}